Add future transaction option and colors
Added checkbox to new transaction form for futures. It sets a BOOL on the backend that is then used to display it differently in the transactions list. We can use it to treat the transactions differently in other instances too.

Also added some basic colors to be used throughout the app. We can change them as needed, but they're centralized in a category and accessible with [UIColor nameOfColor] after importing the UIColor extension.